+++ actrec/__init__.py
@@ -3,7 +3,7 @@
 import bpy
 
 # relative imports
-from . import update, config, preferences, log, operators
+from . import update, config, preferences, log, operators, ar_category, ar_global, ar_local
 # endregion
 
 # region Registration
@@ -11,6 +11,9 @@ def register():
     update.register()
     preferences.register()
     operators.register()
+    ar_local.register()
+    ar_category.register()
+    ar_global.register()
     for cls in ActionRecorder.classes:
         bpy.utils.register_class(cls)
     for cls in ActionRecorder.classespanel:
@@ -21,9 +24,15 @@ def register():
         except:
             continue
     ActionRecorder.Initialize_Props()
-    log_sys.logger.info("Registered Action Recorder")
+    log.logger.info("Registered Action Recorder")
 
 def unregister():
+    update.unregister()
+    preferences.unregister()
+    operators.unregister()
+    ar_local.unregister()
+    ar_category.unregister()
+    ar_global.unregister()
     for cls in ActionRecorder.classes:
         bpy.utils.unregister_class(cls)
     for cls in ActionRecorder.categoriesclasses:
