--- actrec/ar_category/__init__.py
@@ -1,15 +1,0 @@
-# region Imports
-# blender modules
-import bpy
-
-# relativ imports
-from . import operators, panles, preferences, functions
-# endregion
-
-# region Registration
-def register():
-    operators.register()
-
-def unregister():
-    operators.unregister()
-# endregion
