--- actrec/ar_global/__init__.py
@@ -1,7 +1,0 @@
-from . import preferences, functions
-
-def register():
-    preferences.register()
-
-def register():
-    preferences.unregister()
