--- actrec/ar_local/__init__.py
@@ -1,12 +1,0 @@
-# region Imports
-# blender modules
-import bpy
-# endregion
-
-# region Registration
-def register():
-    pass
-
-def unregister():
-    pass
-# endregion
