--- actrec/icon_manager/__init__.py
@@ -1,16 +1,0 @@
-# region Imports
-# blender modules
-import bpy
-
-# relative imports
-from . import operators, functions
-# endregion
-
-
-# region Registration 
-def register() -> None:
-    operators.register()
-
-def unregister() -> None:
-    operators.register()
-# endregion
