+++ actrec/log.py
@@ -26,18 +26,18 @@ class log_system:
                 os.remove(min([os.path.join(dirc, filename) for filename in all_logs], key=os.path.getctime)) # delete oldest file
                 all_logs = os.listdir(dirc)
             except PermissionError as err:
-                loglater.append("File is already used -> PermissionError: " + str(err))
+                loglater.append("File is already used -> PermissionError: %" %str(err))
                 break
         path = os.path.join(dirc, "ActRec_%s.log" % bpy.app.tempdir.split("\\")[-2].split("_")[1]) #get individuell id from blender tempdir
 
-        logger = logging.getLogger(__name__)
+        logger = logging.getLogger(__package__)
         logger.setLevel(logging.INFO)
         file_formatter = logging.Formatter("%(levelname)s - %(relativeCreated)d - %(funcName)s - %(message)s")
         file_handler = logging.FileHandler(path, mode='w', encoding= 'utf-8', delay= True)
         file_handler.setLevel(logger.level)
         file_handler.setFormatter(file_formatter)
         logger.addHandler(file_handler)
-        logger.info("Logging ActRec " + ".".join([str(x) for x in bl_info['version']]) + " running on Blender " + bpy.app.version_string)
+        logger.info("Logging ActRec %s %s %s" %(".".join([str(x) for x in bl_info['version']]), "running on Blender", bpy.app.version_string))
         for log_text in loglater:
             logger.info(log_text)
         self.logger = logger
