+++ actrec/update.py
@@ -1,17 +1,17 @@
 # region Import
 # external modules
-from typing import Optional, Text
+from typing import Optional
 import requests
 import json
 import base64
 import os
 import zipfile
-import sys
 import subprocess
+from collections import defaultdict
 
 # blender modules
 import bpy
-from bpy.types import Operator, AddonPreferences, Menu
+from bpy.types import Operator, AddonPreferences
 from bpy.props import BoolProperty, IntProperty, StringProperty
 from bpy_extras.io_utils import ExportHelper
 
@@ -24,7 +24,7 @@ from .preferences import AR_preferences
 classes = []
 
 # region functions
-def on_start(dummy = None) -> None:
+def on_start() -> None:
     check_for_update()
 
 def get_json_from_content(content : bytes) -> dict:
@@ -54,9 +54,6 @@ def update(AR, update_responds: dict, download_chunks: dict) -> bool:
     complet_length = 0
     complet_progress = 0
     for path in update_responds:
-        download_chunks.setdefault(path, {})
-        download_chunks[path].setdefault("chunks", [])
-        download_chunks[path].setdefault("progress_length", 0)
         res = update_responds[path]
         total_length = res.headers.get('content-length', None)
         complet_length += total_length
@@ -87,7 +84,6 @@ def install_update(AR, download_chunks: dict) -> None:
     os.remove(zip_path)
     update_download_file(download_chunks.keys(), version)
     download_chunks.clear()
-    AR.
 
 def get_online_download_file() -> Optional[dict]:
     try:
@@ -99,11 +95,12 @@ def get_online_download_file() -> Optional[dict]:
         return None
 
 def get_local_download_file() -> dict:
-    if not os.path.exists("download_file.json"):
-        new_file = open("download_file.json", "w", encoding= 'utf-8')
+    path = os.path.join(AR_preferences.addon_directory, "download_file.json")
+    if not os.path.exists(path):
+        new_file = open(path, "w", encoding= 'utf-8')
         new_file.write("{}")
         new_file.close()
-    download_file = open("download_file.json", "r", encoding= 'utf-8')
+    download_file = open(path, "r", encoding= 'utf-8')
     download_text = download_file.read()
     download_file.close()
     return json.loads(download_text)
@@ -126,7 +123,8 @@ def get_download_paths() -> Optional[list]:
     download_list = []
     old_keys = old_download_file.keys()
     for key in download_file.keys():
-        if (key in old_keys and download_file[key] > old_download_file) or not(key in old_keys):
+        key_exists = key in old_keys 
+        if (key_exists and download_file[key] > old_download_file[key]) or not key_exists:
             download_list.append(key)
     return download_list
 
@@ -136,14 +134,16 @@ def draw_update_button(layout, AR):
         row.enable = False
         row.prop(AR, 'update_progress', text= "Progress", slider=True)
     else:
-        layout.operator(AR_OT_Update.bl_idname, text= 'Update')
+        layout.operator(AR_OT_update.bl_idname, text= 'Update')
 # endregion functions
 
 # region Operator
-class AR_OT_CheckUpdate(Operator):
-    bl_idname = "ar.check_update"
+class AR_OT_update_check(Operator):
+    bl_idname = "ar.update_check"
     bl_label = "Check for Update"
     bl_description = "check for available update"
+
+
 
     def execute(self, context):
         update = check_for_update()
@@ -154,9 +154,9 @@ class AR_OT_CheckUpdate(Operator):
         else:
             AR.version = ".".join([str(i) for i in update[1]])
         return {"FINISHED"}
-classes.append(AR_OT_CheckUpdate)
+classes.append(AR_OT_update_check)
 
-class AR_OT_Update(Operator):
+class AR_OT_update(Operator):
     bl_idname = "ar.update"
     bl_label = "Update"
     bl_description = "install the new version"
@@ -164,7 +164,7 @@ class AR_OT_Update(Operator):
     @classmethod
     def poll(cls, context):
         AR = context.preferences.addons[__package__].preferences
-        return AR.Update
+        return AR.update
 
     def invoke(self, context, event):
         launch = start_update()
@@ -187,9 +187,9 @@ class AR_OT_Update(Operator):
         AR.update_progress = -1
         bpy.ops.ar.show_restart_menu('INVOKE_DEFAULT')
         return {"FINISHED"}
-classes.append(AR_OT_Update)
+classes.append(AR_OT_update)
 
-class AR_OT_Restart(Operator, ExportHelper):
+class AR_OT_restart(Operator, ExportHelper):
     bl_idname = "ar.restart"
     bl_label = "Restart Blender"
     bl_description = "Restart Blender"
@@ -199,6 +199,10 @@ class AR_OT_Restart(Operator, ExportHelper):
     filter_folder : BoolProperty(default= True, options={'HIDDEN'})
     filter_blender : BoolProperty(default= True, options={'HIDDEN'})
 
+    @classmethod
+    def poll(cls, context):
+        AR = context.preferences.addons[__package__].preferences
+        return AR.restart
 
     def execute(self, context):
         path = bpy.data.filepath
@@ -216,9 +220,9 @@ class AR_OT_Restart(Operator, ExportHelper):
             return ExportHelper.invoke(self, context, event)
         else:
             return self.execute(context)
-classes.append(AR_OT_Restart)
+classes.append(AR_OT_restart)
 
-class AR_OT_Show_Restart_Menu(Operator):
+class AR_OT_show_restart_menu(Operator):
     bl_idname = "ar.show_restart_menu"
     bl_label = "Restart Blender"
     bl_description = "Restart Blender"
@@ -233,24 +237,24 @@ class AR_OT_Show_Restart_Menu(Operator):
         box = layout.box()
         box.label(text= "You need to restart Blender to complete the Update")
         row = self.layout.row()
-        row.operator(AR_OT_Restart.bl_idname, text= "Save & Restart").save = True
-        row.operator(AR_OT_Restart.bl_idname, text= "Restart")
+        row.operator(AR_OT_restart.bl_idname, text= "Save & Restart").save = True
+        row.operator(AR_OT_restart.bl_idname, text= "Restart")
 
     def execute(self, context):
         context.window_manager.popup_menu(self.draw, title= "Action Recorder Restart")
         return {"FINISHED"}
-classes.append(AR_OT_Show_Restart_Menu)
+classes.append(AR_OT_show_restart_menu)
 # endregion
 
 # region Preferences
-class Preferneces(AddonPreferences):
+class preferences(AddonPreferences):
     launch_update : BoolProperty()
     restart : BoolProperty()
     version : StringProperty()
     auto_update : BoolProperty(default= True, name= "Auto Update", description= "automatically search for a new Update")
     update_progress : IntProperty(name= "Update Progress", default= -1, min= -1, max= 100, soft_min= 0, soft_max= 100, subtype= 'PERCENTAGE') # use as slider
     update_responds = {}
-    update_data_chunks = {}
+    update_data_chunks = defaultdict(lambda: {"chunks": [], 'progress_length': 0})
 # endregion
 
 # region Regestration
